### 2021-09-07 #3697 @whedon
[PRE REVIEW]: epyc: Computational experiment management in Python
**Submitting author:** @simoninireland (<a href="http://orcid.org/0000-0001-9633-2103">Simon Dobson</a>)
**Repository:** <a href="https://github.com/simoninireland/epyc" target ="_blank">https://github.com/simoninireland/epyc</a>
**Version:** 1.5.1
