### 2021-09-07 #3697 @arfon
:wave: @simoninireland – looks like there are some whitespace issues with your `paper.md` YAML header (see how GitHub is rendering here: https://github.com/simoninireland/epyc/blob/master/paper.md). Could you please fix it up and test here with `@whedon generate pdf`
