### 2021-09-07 #3697 @simoninireland
@arfon That would have been Emacs tabify doing its thing... Fixed now :-)
