### 2021-09-17 #3697 @zbeekman
@ajstewartlang Anything I can do to help get this review going? Suggest fellow reviewers etc? I seem to recall that the submitting author (@simoninireland) can also peruse [the list of potential reviewers](https://bit.ly/joss-reviewershttps://bit.ly/joss-reviewers) and suggest them here. In past reviews I've always been invited by an @ mention in the comments.
