### 2021-09-17 #3697 @ajstewartlang
If you could @zbeekman that would be super helpful. I've emailed a couple of possible reviewers but haven't yet had responses - I appreciate people are understandably quite busy at the moment so if you or @simoninireland spotted some potential reviewers in that [list](https://bit.ly/joss-reviewershttps://bit.ly/joss-reviewers), that would be great thanks.
