### 2021-09-25 #3764 @whedon
[REVIEW]: epyc: Computational experiment management in Python
**Submitting author:** @simoninireland (<a href="http://orcid.org/0000-0001-9633-2103">Simon Dobson</a>)
**Repository:** https://github.com/simoninireland/epyc
**Branch with paper.md** (empty if default branch):
**Version:** v1.6.2
