### 2021-09-26 #3764 @zbeekman
Sorry, I think I neglected to accept the GitHub joss invitation in a timely fashion and now it’s expired. Could an editor (@ajstewartlang @arfon etc) please re-invite me so I can edit the issue?

I think the command is:

thanks
