### 2021-09-28 f5ec84a
Added LabNotebook.createWith() for guarded creation
