### 2021-09-28 20d5390
Added LabNotebook.createWith() for guarded creation
