### 2021-10-02 simoninireland/epyc#7 @lorenzo-rovigatti
The `epyc` console script does not work
Hi @simoninireland ,

This is an issue created as part of [epyc's JOSS review](https://github.com/openjournals/joss-reviews/issues/3764).

I have noticed that the setup procedure installs the `epyc` console script. However, the script fails with `ModuleNotFoundError: No module named 'epyc.scripts'`. I have tested both PyPi and github versions.
