### 2021-10-02 simoninireland/epyc#8 @lorenzo-rovigatti
Recommend to use pip rather setup.py
This is an issue created as part of [epyc's JOSS review](https://github.com/openjournals/joss-reviews/issues/3764).

This is a rather minor (and maybe even opinion-based) point, but I think that you should recommend using `pip` to install the package even when cloned from github.
The main advantage of `pip` over `setup.py` (at least on a linux box) is that the former installs the package in the user home folder by default, whereas the latter gives out an error if the user doesn't have the privileges to install the package in a system folder.
