### 2021-10-02 simoninireland/epyc#9 @lorenzo-rovigatti
Testing the installation
This is an issue created as part of [epyc's JOSS review](https://github.com/openjournals/joss-reviews/issues/3764).

I see that the repo has a CI workflow active. However, I don't see any way of triggering the testing suite manually so that I can check whether installation was successful. Can you include the required steps to run the tests in the README and/or in the full documentation?
