### 2021-10-02 simoninireland/epyc#10 @lorenzo-rovigatti
Add examples to the repo
This is an issue created as part of [epyc's JOSS review](https://github.com/openjournals/joss-reviews/issues/3764).

While the online documentation has tutorials that are very well done, I think it would greatly help future users to have an example folder where these tutorials (or similar ones) can be found. In this way the user can immediately start tinkering with the code.

Note that the `doc/epyc.ipynb` notebook, which is a great resource, requires jupyter (which is not required by the package) and there are no instructions about to launch it, or at least I couldn't find them.
