### 2021-10-04 simoninireland/epyc#7 @simoninireland
Always helps if you include the module in setup.py...:-)
