### 2021-10-04 simoninireland/epyc#10 @simoninireland
Great idea about including examples, I'll move them over from the cookbook etc.

The Jupyter notebook needs a small update, and I didn't want to make it too integral to epyc overall, which is why there's no import of jupyter. On the other hand, we do include all the rest of Jupyter, so maybe there's no reason not to.
