### 2021-10-04 simoninireland/epyc#8 @simoninireland
Do you mean moving to setup.cfg rather than setup.py? I've been wondering about that, but never got round to looking in detail. I'll have another look.
