+++ CONTRIBUTING.rst
@@ -1,0 +1,6 @@
+Contributing to epyc
+====================
+
+Please see the documentation page for epyc's contribution policy:
+
+https://epyc.readthedocs.io/en/latest/contributing.html
