### 2021-10-04 fc5e177
Made license information consistent
