+++ setup.py.in
@@ -31,7 +31,7 @@ setup(name='epyc',
       url='http://github.com/simoninireland/epyc',
       author='Simon Dobson',
       author_email='simoninireland@gmail.com',
-      license='License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
+      license='License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
       classifiers=['Development Status :: 4 - Beta',
                    'Intended Audience :: Science/Research',
                    'Intended Audience :: Developers',
