### 2021-10-04 45a8380
Added GPLv3 logo
