### 2021-10-04 dbe4bf8
Added targets to logos
