+++ doc/tutorial/environment.rst
@@ -1,0 +1,11 @@
+.. _environment:
+
+.. currentmodule:: epyc
+
+A repeatable computational environment
+--------------------------------------
+
+Before we dive in, we should adopt best practices and make sure that
+we have a controlled computational environment in which to perform
+experiments.
+See :ref:`epyc-venv` for a thorough discussion of these topics.
