+++ doc/cookbook/epyc-venv.rst
@@ -5,91 +5,114 @@
 Reproducing experiments reliably
 --------------------------------
 
-**Problem**: Over time the versions numbers of different packages you use change as the code is developed. You're
-worried this might affect your code, either by breaking it or by changing its results somehow.
+**Problem**: Over time the versions numbers of different packages you
+use change as the code is developed. You're worried this might affect
+your code, either by breaking it or by changing its results somehow.
 
-**Solution**: This is a real problem with computational science. Fortunately it's fairly easy to address, at least
-at a simple level.
+**Solution**: This is a real problem with computational
+science. Fortunately it's fairly easy to address, at least at a simple
+level.
 
-Python includes a feature called **virtual environments** or **venvs**. A venv is an installation of Python and its libraries
-that's closed-off from any other installations you may have on your machine. Essentially it takes the global
-installation of Python and  throws away anything that's not part of the core distribution. You can "enter" the
-venv and install exactly those packages you want -- and *only* those packages, and with specific version numbers
-if you like -- secure in the knowledge that if the global environment, or another venv, wants diffrent pacxkages
-and version numnbers they won't interfere with you. You can also "freeze" your venv by grabbing alist of packages and
-bversion numbers installed, and then install this exact environment again later -- or indeed elsewhere, on another
+Python includes a feature called **virtual environments** or
+**venvs**. A venv is an installation of Python and its libraries
+that's closed-off from any other installations you may have on your
+machine. Essentially it takes the global installation of Python and
+throws away anything that's not part of the core distribution. You can
+"enter" the venv and install exactly those packages you want -- and
+*only* those packages, and with specific version numbers if you like
+-- secure in the knowledge that if the global environment, or another
+venv, wants diffrent pacxkages and version numnbers they won't
+interfere with you. You can also "freeze" your venv by grabbing a list
+of packages and version numbers installed, and then install this
+exact environment again later -- or indeed elsewhere, on another
 machine.
 
-Let's assume we want to *create* a venv that we've imaginatively named ``venv``. (You can pick any name you like.)
-You create venvs from the command line:
+Let's assume we want to *create* a venv that we've imaginatively named
+``venv``. (You can pick any name you like.)  You create venvs from the
+command line:
 
-.. code-block :: sh
+.. code-block:: sh
 
     python3 -m venv ./venv
 
-We next need to *activate* the environment, making it the "current" one that Python will use. This is again done from the
-command line:
+We next need to *activate* the environment, making it the "current"
+one that Python will use. This is again done from the command line:
 
-.. code-block :: sh
+.. code-block:: sh
 
     . venv/bin/activate
 
-This alters the various include paths, command paths, and other elements to make sure that, when you execute the
-Python interpreter or any of the related tools, it runs the ones in the venv and not any others.
+This alters the various include paths, command paths, and other
+elements to make sure that, when you execute the Python interpreter or
+any of the related tools, it runs the ones in the venv and not any
+others.
 
-We next need to *populate* the venv, that is, add the packages we want. We do this using ``pip`` as normal:
+We next need to *populate* the venv, that is, add the packages we
+want. We do this using ``pip`` as normal:
 
-.. code-block :: sh
+.. code-block:: sh
 
     pip3 install ipython ipyparallel
 
-.. note ::
+.. note::
 
-    In some installations, ``pip`` always refers to the ``pip`` tool of Python 2.7, while Python 3's
-    tool is called ``pip3``. It never hurts when unsure to run ``pip3`` explicitly if you're working with
-    Python 3. Similarly you may find there's a tool called ``python3`` or even ``python3.7`` in your venv.
+    In some installations, ``pip`` always refers to the ``pip`` tool
[…]
