### 2021-10-04 f2f4c34
Added examples and command-line documentation
