+++ doc/tutorial/first.rst
@@ -6,6 +6,11 @@ First tutorial: Getting started
 In the first tutorial we'll look at the basic components of ``epyc``, and how to define and
 run experiments on a single machine.
 
+.. note::
+
+   The code for this tutorial can be found
+   `in the Github repository <https://github.com/simoninireland/epyc/blob/master/doc/examples/first-tutorial/first.py>`_.
+
 .. include:: concepts.rst
 
 .. include:: environment.rst
