### 2021-10-04 e19048f
Added examples and command-line documentation
