### 2021-10-04 e6ae7f8
Changed current sub-command to select to match Python API
