### 2021-11-04 simoninireland/epyc#8 @lorenzo-rovigatti
No, it's a much simpler thing. As of now you recommend installing the package with `python setup.py install`, whereas I think it's simpler to do `pip install .`, so that the package is always managed by pip. As I wrote, it's a very minor thing.
