### 2021-11-04 simoninireland/epyc#10 @lorenzo-rovigatti
The example you added to the `dev` branch does not work on neither branch. I attach here the traceback:

```
Traceback (most recent call last):
  File "first.py", line 64, in <module>
    lab.runExperiment(CurveExperiment())
  File "/home/lorenzo/.local/lib/python3.8/site-packages/epyc/lab.py", line 205, in runExperiment
    nb.addResult(res)
  File "/home/lorenzo/.local/lib/python3.8/site-packages/epyc/labnotebook.py", line 506, in addResult
    self._addResult(results, tag)
  File "/home/lorenzo/.local/lib/python3.8/site-packages/epyc/labnotebook.py", line 470, in _addResult
    rs.addSingleResult(result)
  File "/home/lorenzo/.local/lib/python3.8/site-packages/epyc/resultset.py", line 687, in addSingleResult
    dt = self.inferDtype(rc)
  File "/home/lorenzo/.local/lib/python3.8/site-packages/epyc/resultset.py", line 553, in inferDtype
    self._results = DataFrame(columns=types.keys())
  File "/usr/lib/python3/dist-packages/pandas/core/frame.py", line 411, in __init__
    mgr = init_dict(data, index, columns, dtype=dtype)
  File "/usr/lib/python3/dist-packages/pandas/core/internals/construction.py", line 242, in init_dict
    val = construct_1d_arraylike_from_scalar(np.nan, len(index), nan_dtype)
  File "/usr/lib/python3/dist-packages/pandas/core/dtypes/cast.py", line 1229, in construct_1d_arraylike_from_scalar
    dtype = dtype.dtype
AttributeError: type object 'object' has no attribute 'dtype'
```
