### 2021-11-05 simoninireland/epyc#11 @zbeekman
JOSS-paper DOIs
Hi, it looks like you have a DOI for *Design of comparative experiments* but perhaps the ISBN should be removed to display that instead? [I tried told @whedon to regenerate the draft PDF](https://github.com/openjournals/joss-reviews/issues/3764#issuecomment-962192322) but the generated document still shows the ISBN. Maybe I should ask the JOSS editors if one is preferred over the other.

Also, my initial searching seems to indicate that there is no DOI available for *Epidemic modelling -- some notes, maths and code.* and likewise there is no DOI for the Software Sustainability Institute *Guides for researchers*. Is this correct? Have you searched for DOIs for these items?
