### 2021-11-05 #3764 @zbeekman
@ajstewartlang or @danielskatz I have questions about citations in the bibliography:

1. Are DOIs preferred over ISBNs?
2. Do all references need a DOI?

Relevant to https://github.com/simoninireland/epyc/issues/11
