### 2021-11-05 #3764 @danielskatz
DOIs are preferred and required for references that have them.
