### 2021-11-05 #3764 @zbeekman
@danielskatz the paper already [includes DOIs for both of those references](https://github.com/simoninireland/epyc/blob/136456558e1d48d62d68df43312fcfc739120915/paper.bib#L22-L43), I'm not sure why @whedon is saying they're missing. Maybe there's a formatting/capitalization issue?

Additionally, the bibliography [provides both an ISBN and DOI for Bailey, R. A. (2009). *Design of comparative experiments*](https://github.com/simoninireland/epyc/blob/136456558e1d48d62d68df43312fcfc739120915/paper.bib#L45-L52) but pandoc/the preview system are showing the ISBN rather than the DOI in the draft.
