### 2021-11-05 simoninireland/epyc#11 @zbeekman
It looks like pandoc/the document compilation system is not correctly recognizing your DOIs (https://github.com/openjournals/joss-reviews/issues/3764#issuecomment-962197918). I'm not sure whether this is due to the capitalization of the DOI url in the bibtex `url` field, or if you need to use a different field in the bibtex style bibliography to make it recognize the doi.

It looks like maybe you need to be using the bibtex `doi` field instead of `url`: https://www.bibtex.com/f/doi-field/
