### 2021-11-05 simoninireland/epyc#11 @zbeekman
(Also, [the editor confirmed](https://www.bibtex.com/f/doi-field/) DOIs are preferred over ISBN etc. Perhaps changing the bibtex entries to use the `doi` field will resolve this issue of the isbn being preferred over the doi?)
