### 2021-11-05 #3764 @danielskatz
please update the bib file to store dois in doi = {} fields, not url fields.  these should be of the form doi = {a.b/...} without the full URL listing
