### 2021-11-05 #3764 @zbeekman
@simoninireland the paper looks great! I opened two issues on the project/paper repository. The first is to address/fix the reference DOI display. The second is that there is an unclosed parenthetical clause at the end of the "Differences with other tools" section. Once those are resolved I can check off all the boxes for the paper.

I'm now moving on to examine the documentation and verify the installation and functionality now. I apologize once again that I haven't gotten through my review sooner. I can't seem to catch a break in my personal/family life: We had to say goodbye to our beloved [family dog](https://github.com/zbeekman/TheodoreVonKarmanTheGreyt) this week. (After a few busy days of taking him to the veterinary hospital, scheduling appointments & picking up medications as his condition worsened.) Thanks so much for your patience and for the interesting tool and paper--it's helped distract me a bit from my recent loss.
