### 2021-11-05 simoninireland/epyc#11 @zbeekman
> please update the bib file to store dois in doi = {} fields, not url fields. these should be of the form doi = {a.b/...} without the full URL listing

I'll submit a PR in a bit if I have time
