+++ setup.py
@@ -25,13 +25,13 @@ with open('README.rst') as f:
     longDescription = f.read()
 
 setup(name='epyc',
-      version='1.5.1',
+      version='1.6.1',
       description='Python computational experiment management',
       long_description=longDescription,
       url='http://github.com/simoninireland/epyc',
       author='Simon Dobson',
       author_email='simoninireland@gmail.com',
-      license='License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
+      license='License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
       classifiers=['Development Status :: 4 - Beta',
                    'Intended Audience :: Science/Research',
                    'Intended Audience :: Developers',
@@ -41,7 +41,9 @@ setup(name='epyc',
                    'Programming Language :: Python :: 3.9',
                    'Topic :: Scientific/Engineering'],
       python_requires='>=3.5',
-      packages=['epyc'],
+      packages=['epyc',
+                'epyc.scripts',
+                ],
       include_package_data=True,
       package_data={'epyc': ['py.typed']},
       entry_points='''
@@ -51,4 +53,3 @@ setup(name='epyc',
       zip_safe=False,
       install_requires=["numpy >= 1.17.5", "pyzmq", "ipyparallel >= 6.2.4", "cloudpickle", "pandas", "h5py", "joblib", "requests", "click", ],
       extra_requires={':python_version < 3.8': ['typing_extensions']})
-
