+++ tox.ini
@@ -3,4 +3,4 @@ envlist = py39
 
 [testenv]
 deps = -r dev-requirements.txt
-commands = nosetests []
+commands = python -m unittest discover
