+++ Makefile
@@ -130,7 +130,8 @@ SOURCES_UTILS = \
 SOURCES_EXTRA = \
 	README.rst \
 	LICENSE \
-	HISTORY
+	HISTORY \
+	CONTRIBUTING.rst
 SOURCES_GENERATED = \
 	MANIFEST \
 	setup.py
