### 2021-11-12 0a6aefa
Refactored createWith to Lab instead of LabNotebook
