+++ test/test_notebooks.py
@@ -24,18 +24,6 @@ from datetime import datetime
 
 class LabNotebookTests(unittest.TestCase):
 
-    def create(self):
-        '''Create some results into the current result set.'''
-        rc1 = self._resultsdict()
-        rc1[Experiment.PARAMETERS]['a'] = 1
-        rc1[Experiment.PARAMETERS]['b'] = 2
-        rc1[Experiment.RESULTS]['first'] = 3
-        self._nb.addResult(rc1)
-
-    def createFail(self):
-        '''Faail to create results.'''
-        raise Exception('Failed!')
-
     def setUp(self):
         '''Set up a notebook.'''
         self._nb = LabNotebook()
@@ -339,11 +327,12 @@ class LabNotebookTests(unittest.TestCase):
 
     def testDeleteResultSetByTag(self):
         '''Test we can delete a result set.'''
+        rs = self._nb.addResultSet('first')
         rs = self._nb.addResultSet('second')
-        self._nb.deleteResultSet(LabNotebook.DEFAULT_RESULTSET)
-        self.assertEqual(len(self._nb), 1)
+        self._nb.deleteResultSet('first')
+        self.assertEqual(len(self._nb), 2)
         self.assertEqual(self._nb.currentTag(), 'second')
-        self.assertCountEqual(self._nb.resultSets(), ['second'])
+        self.assertCountEqual(self._nb.resultSets(), ['second', LabNotebook.DEFAULT_RESULTSET])
         self.assertEqual(self._nb.current(), rs)
 
     def testDeleteResultSetByReference(self):
@@ -360,6 +349,11 @@ class LabNotebookTests(unittest.TestCase):
         rs = self._nb.addResultSet('second')
         with self.assertRaises(Exception):
             self._nb.deleteResultSet(rs)
+
+    def testDeleteDefault(self):
+        '''Test we can't delete the default result set.'''
+        with self.assertRaises(Exception):
+            self._nb.deleteResultSet(LabNotebook.DEFAULT_RESULTSET)
 
     def testKeys(self):
         '''Test we can extract the result set tags.'''
@@ -438,47 +432,6 @@ class LabNotebookTests(unittest.TestCase):
 
         p = self._nb.pendingResultParameters('1234')
         self.assertCountEqual(rc1[Experiment.PARAMETERS], p)
-
-    def testBasicCreateWith(self):
-        '''Test createWith works.'''
-        rc = self._nb.createWith('test1', self.create, description='A first set')
-        self.assertTrue(rc)
-        self.assertEqual(self._nb.current().description(), 'A first set')
-        self.assertIn('test1', self._nb.resultSets())
-        self.assertFalse(self._nb.current().isLocked())
-
-    def testCreateWithFinishes(self):
-        '''Test createWith finishes a result set when requested.'''
-        rc = self._nb.createWith('test1', self.create, finish=True)
-        self.assertTrue(rc)
-        self.assertTrue(self._nb.current().isLocked())
-
-    def testCreateWithFail(self):
-        '''Test createWith handles deletion on failure.'''
-        with self.assertRaises(Exception):
-            rc = self._nb.createWith('test1', self.createFail)
-        self.assertNotIn('test1', self._nb.resultSets())
-
-    def testCreateWithFailNoDelete(self):
-        '''Test createWith can be prevented from deleting on failure.'''
-        with self.assertRaises(Exception):
-            rc = self._nb.createWith('test1', self.createFail, delete=False)
-        self.assertIn('test1', self._nb.resultSets())
-        self.assertFalse(self._nb.current().isLocked())
-
-    def testCreateWithFailNoDeleteNotFinished(self):
-        '''Test createWith doesn't lock a partial result set.'''
-        with self.assertRaises(Exception):
[…]
