### 2021-11-12 4d98f90
Activate before removing PID file
