### 2021-11-12 67062bd
Updated to new matplotlib 3d syntax
