### 2021-11-12 56a6832
Updated to new matplotlib 3d syntax
