+++ doc/tutorial/results.rst
@@ -54,7 +54,7 @@ cloud:
     import matplotlib.pyplot as plt
 
     fig = plt.figure(figsize = (8, 8))
-    ax = fig.gca(projection = '3d')
+    ax = fig.add_subplot(projection = '3d')
 
     ax.scatter(df['x'], df['y'], df['result'],
 	       c=df['result'], depthshade=False, cmap=cm.coolwarm)
