### 2021-11-12 simoninireland/epyc#10 @simoninireland
Fixed (I think) -- and I took the opportunity to update the matplotlib code to the new 3d interface as well
