### 2021-11-12 simoninireland/epyc#11 @simoninireland
Updated to DOI fields. The bibtex style seems to prefer ISBNs to DOIs, so I removed the ISBN from "Design of comparative experiments".

There is no DOI for "Epidemic modelling" (I know because I wrote it :-)), and none for the SSI's "Guide for researchers" as it's purely online as a web page.
