### 2021-11-12 simoninireland/epyc#9 @simoninireland
The test suite isn't part of the package (as is normally the case, I believe), so it isn't installed when you install the package. You can of course run tests if you have a repo checked out: make test will do that.
