+++ paper.bib
@@ -5,7 +5,6 @@ publisher = "Independent Publishing Network",
 year = 2020,
 isbn = "978-183853-565-0",
 url = "https://simoninireland.github.io/introduction-to-epidemics/",
-keywords = "epidemic spreading, complex networks, disease, coronavirus, covid-19, epydemic, Python",
 }
 
 @article{SymbioticAntagonisticClustered,
@@ -15,8 +14,7 @@ journal = "Physical Review E",
 volume = 104,
 number = 2,
 year = 2021,
-url = "https://doi.org/10.1103/PhysRevE.104.024303",
-keywords = "complex networks, percolation theory, epidemic spreading, co-infection, clustering",
+doi = "10.1103/PhysRevE.104.024303",
 }
 
 @article{ClusteredTwoPathogen,
@@ -27,8 +25,7 @@ volume = 103,
 number = 6,
 month = "June",
 year = 2021,
-url = "https://doi.org/10.1103/PhysRevE.103.062308",
-keywords = "complex networks, percolation theory, epidemic spreading, co-infection, clustering",
+doi = "10.1103/PhysRevE.103.062308",
 }
 
 @article{TB,
@@ -39,7 +36,7 @@ volume = 506,
 month = "June",
 year = 2020,
 keywords = "tuberculosis, disease, complex networks, simulation",
-url = "https://dx.doi.org//10.1101/2019.12.12.871269 ",
+doi = "10.1101/2019.12.12.871269",
 }
 
 @book{DesignComparativeExperiments,
@@ -47,9 +44,9 @@ author = "R.A. Bailey",
 title = "Design of comparative experiments",
 publisher = "Cambridge University Press",
 year = 2009,
-isbn = "978-051161-148-3",
-url = "https://doi.org/10.1017/CBO9780511611483",
+doi = "10.1017/CBO9780511611483",
 }
+%isbn = "978-051161-148-3",
 
 @misc{SSI,
 author = "{{S}oftware {S}ustainability {I}nstitute}",
