### 2021-11-26 simoninireland/epyc#9 @lorenzo-rovigatti
OK thanks. I have tried using `make test` but it fails with the following error (ubuntu 20.04):

```
cat setup.py.in | sed -e 's|VERSION|1.5.1|g' -e 's|REQUIREMENTS|"numpy >= 1.17.5", "pyzmq", "ipyparallel >= 6.2.4", "cloudpickle", "pandas", "h5py", "joblib", "requests", "click", |g' >setup.py
. venv3/bin/activate && tox
/bin/sh: 1: .: Can't open venv3/bin/activate
make: *** [Makefile:203: test] Errore 127
```
