### 2021-11-26 #3764 @lorenzo-rovigatti
Hey! It's mostly there for me. There is the small issue about testing (which is what the repo's issue is about) and then the paper is good to go as soon as I am concerned.
