### 2022-01-11 simoninireland/epyc#9 @ajstewartlang
@simoninireland do you have any thoughts re. why that error highlighted by @lorenzo-rovigatti occurs? Is there another way to test whether installation is successful?
