### 2022-01-18 simoninireland/epyc#9 @simoninireland
That's the venv not having been built, which is weird if done from scratch as env is a dependency of test. Running "make reallyclean && make test" should fix whatever's happening.
