### 2022-01-18 9c033f6
Typos and minor layout changes
