+++ epyc/repeatedexperiment.py
@@ -1,6 +1,6 @@
 # Experiment combinator to run the same experiment several times.
 #
-# Copyright (C) 2016--2020 Simon Dobson
+# Copyright (C) 2016--2022 Simon Dobson
 #
 # This file is part of epyc, experiment management in Python.
 #
@@ -17,6 +17,7 @@
 # You should have received a copy of the GNU General Public License
 # along with epyc. If not, see <http://www.gnu.org/licenses/gpl.html>.
 
+from epyc import ExperimentCombinator, Experiment, ResultsDict
 import sys
 from typing import Any, Dict, List, Union
 if sys.version_info >= (3, 8):
@@ -24,7 +25,6 @@ if sys.version_info >= (3, 8):
 else:
     # backwards compatibility with Python 35, Python36, and Python37
     from typing_extensions import Final
-from epyc import ExperimentCombinator, Experiment, ResultsDict
 
 
 class RepeatedExperiment(ExperimentCombinator):
@@ -42,22 +42,22 @@ class RepeatedExperiment(ExperimentCombinator):
     :pamam N: the number of repetitions to perform"""
 
     # Extra metadata
-    REPETITIONS : Final[str] = 'epyc.experiment.repetitions'   #: Metadata element for number of repetitions performed.
-    I           : Final[str] = 'epyc.experiment.rep'           #: Metadata element for index in the repetitions.
+    REPETITIONS: Final[str] = 'epyc.experiment.repetitions'   #: Metadata element for number of repetitions performed.
+    I          : Final[str] = 'epyc.experiment.rep'           #: Metadata element for index in the repetitions.
 
 
-    def __init__( self, ex : Experiment, N : int ):
-        super(RepeatedExperiment, self).__init__(ex)
+    def __init__(self, ex: Experiment, N: int ):
+        super().__init__(ex)
         self._N = N
 
-    def repetitions( self ) -> int:
+    def repetitions(self) -> int:
         """Return the number of repetitions of the underlying experiment
         we expect to perform.
 
         :returns: the number of repetitions"""
         return self._N
 
-    def do( self, params : Dict[str, Any] ) -> List[ResultsDict]:
+    def do(self, params: Dict[str, Any]) -> List[ResultsDict]:
         """Perform the number of repetitions we want. The results returned
         will be a list of the results dicts generated by the repeated experiments.
         The metadata for each experiment will include an entry
@@ -76,7 +76,7 @@ class RepeatedExperiment(ExperimentCombinator):
 
             # make sure we have a list to traverse
             if not isinstance(rcs, list):
-                rcs = [ rcs ]
+                rcs = [rcs]
 
             # add repetition information to each results dict's metadata
             for j in range(len(rcs)):
