### 2022-01-18 3fb5ca3
Typos and minor layout changes
