+++ epyc/summaryexperiment.py
@@ -1,6 +1,7 @@
+# SUmmary experiment combinator
 #
-# Copyright (C) 2016--2020 Simon Dobson
-# 
+# Copyright (C) 2016--2022 Simon Dobson
+#
 # This file is part of epyc, experiment management in Python.
 #
 # epyc is free software: you can redistribute it and/or modify
@@ -16,14 +17,14 @@
 # You should have received a copy of the GNU General Public License
 # along with epyc. If not, see <http://www.gnu.org/licenses/gpl.html>.
 
-from epyc import ExperimentCombinator, Experiment, ResultsDict
 import numpy                     # type: ignore
+from epyc import ExperimentCombinator, Experiment, ResultsDict
 import sys
+from typing import List, Dict, Any
 if sys.version_info >= (3, 8):
-    from typing import List, Dict, Any, Final
+    from typing import Final
 else:
-    # backwards compatibility with Python 35, Python36, and Python37 
-    from typing import List, Dict, Any
+    # backwards compatibility with Python 35, Python36, and Python37
     from typing_extensions import Final
 
 
@@ -50,18 +51,18 @@ class SummaryExperiment(ExperimentCombinator):
     that is that have their status set to True. Failed runs are ignored."""
 
     # Additional metadata
-    UNDERLYING_RESULTS : Final[str] = 'epyc.summaryexperiment.repetitions'                          #: Metadata element for the number of results that were obtained.
-    UNDERLYING_SUCCESSFUL_RESULTS : Final[str] = 'epyc.summaryexperiment.successful_repetitions'    #: Metadata elements for the number of results that were summarised.
+    UNDERLYING_RESULTS: Final[str] = 'epyc.summaryexperiment.repetitions'                          #: Metadata element for the number of results that were obtained.
+    UNDERLYING_SUCCESSFUL_RESULTS: Final[str] = 'epyc.summaryexperiment.successful_repetitions'    #: Metadata elements for the number of results that were summarised.
 
     # Prefix and suffix tags attached to summarised result and metadata values
-    MEAN_SUFFIX : Final[str] = '_mean'              #: Suffix for the mean of the underlying values.
-    MEDIAN_SUFFIX : Final[str] = '_median'          #: Suffix for the median of the underlying values.
-    VARIANCE_SUFFIX : Final[str] = '_variance'      #: Suffix for the variance of the underlying values.
-    MIN_SUFFIX : Final[str] = '_min'                #: Suffix for the minimum of the underlying values.
-    MAX_SUFFIX : Final[str] = '_max'                #: Suffix for the maximum of the underlying values.
-    
-    
-    def __init__(self, ex : Experiment, summarised_results : List[str] =None):
+    MEAN_SUFFIX: Final[str] = '_mean'              #: Suffix for the mean of the underlying values.
+    MEDIAN_SUFFIX: Final[str] = '_median'          #: Suffix for the median of the underlying values.
+    VARIANCE_SUFFIX: Final[str] = '_variance'      #: Suffix for the variance of the underlying values.
+    MIN_SUFFIX: Final[str] = '_min'                #: Suffix for the minimum of the underlying values.
+    MAX_SUFFIX: Final[str] = '_max'                #: Suffix for the maximum of the underlying values.
+
+
+    def __init__(self, ex: Experiment, summarised_results: List[str] = None):
         """Create a summarised version of the given experiment. The given
         fields in the experimental results will be summarised, defaulting to all.
         If there are fields that can't be summarised (because they're not
@@ -69,30 +70,30 @@ class SummaryExperiment(ExperimentCombinator):
 
         :param ex: the underlying experiment
         :param summarised_results: list of result values to summarise (defaults to all)"""
-        super(SummaryExperiment, self).__init__(ex)
+        super().__init__(ex)
         self._summarised_results = summarised_results
[…]
