### 2022-01-18 5f89a19
Added Python10 compatibility tag
