+++ setup.py.in
@@ -39,6 +39,7 @@ setup(name='epyc',
                    'Programming Language :: Python :: 3.7',
                    'Programming Language :: Python :: 3.8',
                    'Programming Language :: Python :: 3.9',
+                   'Programming Language :: Python :: 3.10',
                    'Topic :: Scientific/Engineering'],
       python_requires='>=3.5',
       packages=['epyc',
