### 2022-01-18 b4102b0
Added logging
