+++ epyc/labnotebook.py
@@ -1,6 +1,6 @@
 # Simulation "lab notebook" for collecting results, in-memory version
 #
-# Copyright (C) 2016--2021 Simon Dobson
+# Copyright (C) 2016--2022 Simon Dobson
 #
 # This file is part of epyc, experiment management in Python.
 #
@@ -17,9 +17,10 @@
 # You should have received a copy of the GNU General Public License
 # along with epyc. If not, see <http://www.gnu.org/licenses/gpl.html>.
 
-from epyc import Experiment, ResultSet, ResultsDict
 from pandas import DataFrame                               # type: ignore
 from contextlib import contextmanager
+import logging
+from epyc import Logger, Experiment, ResultSet, ResultsDict
 import sys
 from typing import List, Set, Dict, Any, Optional, Union, Callable, cast
 if sys.version_info >= (3, 8):
@@ -28,14 +29,17 @@ else:
     from typing_extensions import Final
 
 
+logger = logging.getLogger(Logger)
+
+
 class ResultsStructureException(Exception):
     '''An exception raised when there is a problem with the structure of a
     results dict.
 
     :param rc: the results dict structure that causes the problem'''
 
-    def __init__(self, rc : Union[ResultsDict, List[ResultsDict]]):
-        super().__init__("Can't handle results dict {rc}".format(rc=rc))
+    def __init__(self, rc: Union[ResultsDict, List[ResultsDict]]):
+        super().__init__(f'Can\'t handle results dict {rc}')
         self._rc = rc
 
     def resultsdict(self) -> Union[ResultsDict, List[ResultsDict]]:
@@ -52,8 +56,8 @@ class NotebookVersionException(Exception):
     :param expected: the expected version
     :param actual: the actual version'''
 
-    def __init__(self, expected : str, actual : str):
-        super().__init__('Expected notebook version {v1}, got {v2}'.format(v1=expected, v2=actual))
+    def __init__(self, expected: str, actual: str):
+        super().__init__(f'Expected notebook version {expected}, got {actual}')
         self._expected = expected
         self._actual = actual
 
@@ -101,21 +105,21 @@ class LabNotebook:
     '''
 
     # Defaults
-    DEFAULT_RESULTSET : Final[str] = 'epyc.resultset.default'  #: Tag for the default result set.
+    DEFAULT_RESULTSET: Final[str] = 'epyc.resultset.default'  #: Tag for the default result set.
 
-    def __init__(self, name : str ='', description : str =None):
+    def __init__(self, name: str='', description: str = None):
         if description is None:
             description = 'A notebook'
-        self._name : Optional[str] = name                    # name
-        self._description : str = description                # description
-        self._resultSets : Dict[str, ResultSet] = dict()     # tag to result set
-        self._resultSetTags : Dict[ResultSet, str] = dict()  # result set to tag
-        self._pending : Dict[str, ResultSet] = dict()        # pending results job ids to result sets
-        self._locked : bool = False                          # locked flag
+        self._name: Optional[str] = name                    # name
+        self._description: str = description                # description
+        self._resultSets: Dict[str, ResultSet] = dict()     # tag to result set
+        self._resultSetTags: Dict[ResultSet, str] = dict()  # result set to tag
+        self._pending: Dict[str, ResultSet] = dict()        # pending results job ids to result sets
+        self._locked: bool = False                          # locked flag
 
         # add a result set with the default tag, and make it current
         defrc = self.addResultSet(self.DEFAULT_RESULTSET)
-        self._current : ResultSet = defrc
+        self._current: ResultSet = defrc
         #defrc.dirty(False)                  # default shouldn't trigger a write
 
 
@@ -165,7 +169,7 @@ class LabNotebook:
 
     # ---------- Managing results sets ----------
 
[…]
