### 2022-01-18 6283232
Added logging
