+++ epyc/clusterlab.py
@@ -1,6 +1,6 @@
 # Simulation "lab" experiment management, parallel cluster version
 #
-# Copyright (C) 2016--2020 Simon Dobson
+# Copyright (C) 2016--2022 Simon Dobson
 #
 # This file is part of epyc, experiment management in Python.
 #
@@ -17,12 +17,16 @@
 # You should have received a copy of the GNU General Public License
 # along with epyc. If not, see <http://www.gnu.org/licenses/gpl.html>.
 
-from epyc import Lab, LabNotebook, Experiment
-import numpy                                  # type: ignore
 import time
 import sys
+import logging
 from ipyparallel import Client, DirectView    # type: ignore
 from contextlib import AbstractContextManager
+from epyc import Logger, Lab, LabNotebook, Experiment
+
+
+logger = logging.getLogger(Logger)
+
 
 class ClusterLab(Lab):
     """A :class:`Lab` running on an ``pyparallel`` compute cluster.
@@ -40,11 +44,15 @@ class ClusterLab(Lab):
     """
 
     # Tuning parameters
-    WaitingTime : int = 30           #: Waiting time for checking for job completion. Lower values increase network traffic.
-    Reconnections : int = 5          #: Number of attempts when re-connecting to a cluster.
-    Retries : int = 3                #: Number of re-tries for failed jobs.
+    WaitingTime: int = 30           #: Waiting time for checking for job completion. Lower values increase network traffic.
+    Reconnections: int = 5          #: Number of attempts when re-connecting to a cluster.
+    Retries: int = 3                #: Number of re-tries for failed jobs.
 
-    def __init__( self, notebook : LabNotebook = None, url_file = None, profile = None, profile_dir = None, ipython_dir = None, context = None, debug = False, sshserver = None, sshkey = None, password = None, paramiko = None, timeout = 10, cluster_id = None, **extra_args ):
+    def __init__(self, notebook: LabNotebook = None,
+                 url_file = None, profile = None, profile_dir = None, ipython_dir = None,
+                 context = None, debug = False,
+                 sshserver = None, sshkey = None, password = None, paramiko = None,
+                 timeout = 10, cluster_id = None, **extra_args ):
         """Create an empty lab attached to the given cluster. Most of the arguments
         are as expected by the ``pyparallel.Client`` class, and are used to create the
         underlying connection to the cluster. The connection is opened immediately,
@@ -63,23 +71,23 @@ class ClusterLab(Lab):
         :param paramiko: True to use paramiko for ssh (defaults to False)
         :param timeout: timeout in seconds for ssh connection (defaults to 10s)
         :param cluster_id: string added to runtime files to prevent collisions"""
-        super(ClusterLab, self).__init__(notebook)
+        super().__init__(notebook)
 
         # record all the connection arguments for later
-        self._arguments = dict(url_file = url_file,
-                               profile = profile,
-                               profile_dir = profile_dir,
-                               ipython_dir = ipython_dir,
-                               context = context,
-                               debug = debug,
-                               sshserver = sshserver,
-                               sshkey = sshkey,
-                               password = password,
-                               paramiko = paramiko,
-                               timeout = timeout,
-                               cluster_id = cluster_id,
+        self._arguments = dict(url_file=url_file,
+                               profile=profile,
+                               profile_dir=profile_dir,
+                               ipython_dir=ipython_dir,
+                               context=context,
+                               debug=debug,
+                               sshserver=sshserver,
+                               sshkey=sshkey,
+                               password=password,
+                               paramiko=paramiko,
+                               timeout=timeout,
+                               cluster_id=cluster_id,
                                **extra_args)
-        self._client : Client = None
+        self._client: Client = None
 
         # connect to the cluster
         self.open()
[…]
