### 2022-01-18 da2d03b
Removed explicit logging setup
