### 2022-01-18 be352e9
Added recipe on logging
