+++ doc/cookbook/metadata.rst
@@ -1,6 +1,6 @@
 .. _resultset-metadata:
 
-.. currentmodule :: epyc
+.. currentmodule:: epyc
 
 Making data archives
 --------------------
@@ -16,7 +16,7 @@ both in technical and semantic terms.
 
 The technical part -- a file that's in an outdated format -- is the easier problem
 to deal with. We can use a format that's already survived the test of time,
-that has widespread support, and that -- althoughbit eventually *will* go out
+that has widespread support, and that -- although it eventually *will* go out
 of date -- will have emough commitment that it'll be possible to convert and
 upgrade it. HDF5, as used by the :class:`HDF5LabNotebook`, meets these criteria
 well, and can be accessed natively by ``epyc``.
@@ -24,7 +24,7 @@ well, and can be accessed natively by ``epyc``.
 Note that ``epyc`` also records the class names of experiments in their results.
 This is only a guide, of course: there's nothing that automatically identifies where
 the code of a class is stored, or which version was used. It's possible to address
-these issues as part of dataset semantics, though. 
+these issues as part of dataset semantics, though.
 
 The semantic problem requires that we maintain an understanding of what each field
 in a dataset *means*. At a trivial level, sensible field names help, as do free-text
@@ -36,27 +36,27 @@ One can be even more structured. Each parameter and result field in a result set
 each metadata field, for that matter) will presumably have a particular purpose and
 likely some units. We can use attributes to store this metadata too:
 
-.. code-block :: python
+.. code-block:: python
 
     from epyc import HDF5LabNotebook
 
     # load the notebook and give it a new description
     with HDF5LabNotebook('my-important-dataset.h5') as nb:
-        # set the description
-        nb.setDescription('A notebook I want to understand later')
+	# set the description
+	nb.setDescription('A notebook I want to understand later')
 
-        # select the result set we want to annotate with metadata
-        rs = nb.select('first-experiment')
-        rs.setDescription('Some physics stuff')
+	# select the result set we want to annotate with metadata
+	rs = nb.select('first-experiment')
+	rs.setDescription('Some physics stuff')
 
-        # create attributes for each parameter and result
-        rs[MyExperiment.VELOCITY] = 'Velocity of particle (ms^-1)'
-        rs[MyExperiment.MASS] = 'Mass of particle (g)'
-        rs[MyExperiment.NPARTICLES] = 'Number of particls (number)'
-        rs[MyExperiment.DENSITY] = 'Final particle density (m^-2)'
+	# create attributes for each parameter and result
+	rs[MyExperiment.VELOCITY] = 'Velocity of particle (ms^-1)'
+	rs[MyExperiment.MASS] = 'Mass of particle (g)'
+	rs[MyExperiment.NPARTICLES] = 'Number of particls (number)'
+	rs[MyExperiment.DENSITY] = 'Final particle density (m^-2)'
 
-        # lock the result set against further updates
-        rs.finish()
+	# lock the result set against further updates
+	rs.finish()
 
 We've assumed we have a class ``MyExperiment`` that defines field names for its
 parameter and result fields. For each of these we create an attribute of the result
@@ -67,7 +67,7 @@ would require a far more sophisticated system using ontologies to describe the s
 of information. But it's a start to have the information recorded in a human-readable form,
 closely associated with the data.
 
-In particular fields it may also be worth adhering to specific standards for metadata.
+In particular application domains it may also be worth adhering to specific standards for metadata.
 The UK Digital Curation Centre maintains a `list <https://www.dcc.ac.uk/guidance/standards/metadata/list>`_
 that may be useful.
 
@@ -75,4 +75,3 @@ Finally, we called :meth:`ResultSet.finish` to finish and lock the result set. T
[…]
