### 2022-01-18 e5c5257
Fixed test skipping to match new ipyparallel files
