+++ .github/workflows/ci.yaml
@@ -3,12 +3,13 @@ on:
   push:
     branches:
       - master
+      - dev
 jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8, 3.9]
+        python-version: [3.6, 3.7, 3.8, 3.9, 3.10]
 
     steps:
     - uses: actions/checkout@v2
