### 2022-01-18 177209d
Changed to test discovery rather than nose
