+++ setup.py.in
@@ -2,7 +2,7 @@
 
 # Setup for epyc
 #
-# Copyright (C) 2016--2021 Simon Dobson
+# Copyright (C) 2016--2022 Simon Dobson
 #
 # This file is part of epyc, experiment management in Python.
 #
