### 2022-01-24 simoninireland/epyc#9 @ajstewartlang
@lorenzo-rovigatti can you try @simoninireland's suggestion and see how it goes?
