### 2022-01-24 simoninireland/epyc#9 @lorenzo-rovigatti
Yup, that seems to work. Maybe you could add that line somewhere in the docs? Either way, I think we are good to go!
