+++ tox.ini
@@ -2,5 +2,5 @@
 envlist = py310
 
 [testenv]
-deps = -r requirements.txt
+deps = -r dev-requirements.txt
 commands = python -m unittest discover
