+++ doc/clusterlab.rst
@@ -1,8 +1,10 @@
+.. _clusterlab-class:
+
 :class:`ClusterLab`: Flexible, parallel, asynchronous experiments
 =================================================================
 
 .. currentmodule :: epyc
-   
+
 .. autoclass :: ClusterLab
 
 
@@ -18,12 +20,12 @@ to be done before running experiments, to make all the code used by an
 experiment available in the cluster.
 
 .. automethod :: ClusterLab.numberOfEngines
-		
+
 .. automethod :: ClusterLab.engines
 
 .. automethod :: ClusterLab.sync_imports
 
-		
+
 Running experiments
 -------------------
 
@@ -41,7 +43,7 @@ results that are ready for retrieval, *i.e.*, the fraction of the
 parameter space that has been explored. :meth:`ClusterLab.ready` tests
 whether *all* results are ready. For cases where it is needed (which
 will hopefully be few and far between), :meth:`ClusterLab.wait` blocks
-until all results are ready.  
+until all results are ready.
 
 .. automethod :: ClusterLab.readyFraction
 
@@ -55,13 +57,13 @@ Results management
 
 A cluster lab is performing computation remotely to itself, typically on another machine
 or machines. This means that pending results may become ready spontaneously (from the
-lab's perspective.) Most of the operations that access results first synchronise the 
+lab's perspective.) Most of the operations that access results first synchronise the
 lab's notebook with the cluster, retrieving any results that have been resolved since
 the previous check. (Checks can also be carried out directly.)
 
 .. automethod :: ClusterLab.updateResults
-		
-		
+
+
 Connection management
 ---------------------
 
@@ -77,7 +79,7 @@ cause no problems, as it re-opens automatically when needed.
     there will seldom be a need to use any these methods directly.
 
 .. automethod :: ClusterLab.open
-		
+
 .. automethod :: ClusterLab.close
 
 In a very small number of circumstances it may be necessary to take control
@@ -100,4 +102,3 @@ particular circumstances.
 .. autoattribute :: ClusterLab.Reconnections
 
 .. autoattribute :: ClusterLab.Retries
-
