### 2022-02-27 #3764 @ajstewartlang
@amritagos can you try running the command below please? If it works, it will re-generate your checklist for you and you should then be able to edit the checklist.

`@editorialbot generate my checklist`
