### 2022-02-27 21ac4b1
docs: Minor typographical error
