### 2022-02-27 simoninireland/epyc#14 @amritagos
[JOSS] Manuscript Review
Part of [openjournals/joss-reviews#3764](https://github.com/openjournals/joss-reviews/issues/3764)

The manuscript is succinct, well-written and easy to understand. However, it would be good if the author could provide references for the other workflow managers mentioned in the section on _Differences with other tools_, such as [Snakemake](https://snakemake.readthedocs.io/en/stable/) or [Sacred](https://sacred.readthedocs.io/en/stable/quickstart.html).
