### 2022-02-27 #3764 @amritagos
The package is nicely documented, and is suitable for running repeated computer experiments without resorting to overly complex workflows. The manuscript is well-written and clearly demonstrates the need for `epyc`. I've opened one minor issue, but I'm happy to recommend this for publication.
