### 2022-02-28 #3764 @ajstewartlang
@simoninirelland could you perhaps merge [this PR](https://github.com/simoninireland/epyc/pull/13) and add the references mentioned by @amritagos [here](https://github.com/simoninireland/epyc/issues/14) please? Then I'm happy to click the `accept` button. Thanks for your patience with this and many many thanks to @lorenzo-rovigatti and @amritagos for their time reviewing this great submission!
