### 2022-03-24 55c782c
Added references to Sacred and SnakeMake
