+++ paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
   - name: School of Computer Science, University of St Andrews, Scotland UK
     index: 1
-date: 4 September 2021
+date: 24 March 2022
 bibliography: paper.bib
 ---
 
