+++ Makefile
@@ -1,6 +1,6 @@
 # Makefile for epyc
 #
-# Copyright (C) 2016--2021 Simon Dobson
+# Copyright (C) 2016--2022 Simon Dobson
 #
 # This file is part of epyc, experiment management in Python.
 #
@@ -21,7 +21,7 @@
 PACKAGENAME = epyc
 
 # The version we're building
-VERSION = 1.6.1
+VERSION = 1.6.2
 
 
 # ----- Sources -----
@@ -236,7 +236,7 @@ $(VENV):
 	$(VIRTUALENV) $(VENV)
 	$(CAT) $(REQUIREMENTS) $(DEV_REQUIREMENTS) >$(VENV)/requirements.txt
 	$(ACTIVATE) && $(PIP) install -U pip wheel && $(CHDIR) $(VENV) && $(PIP) install -r requirements.txt
-	$(ACTIVATE) && mypy --install-types --non-interactive
+	$(ACTIVATE) && mypy --install-types --non-interactive epyc/ --exclude epyc/scripts
 
 # Build a source distribution
 sdist: $(DIST_SDIST)
