+++ CITATION.cff
@@ -1,0 +1,10 @@
+cff-version: 1.2.0
+message: "If you use this software and want to cite it, please reference the following."
+authors:
+- family-names: "Dobson"
+  given-names: "Simon"
+  orcid: "https://orcid.org/0000-0001-9633-2103"
+title: "epyc"
+doi: 10.5281/zenodo.6460760
+date-released: 2022-04-14
+url: "https://github.com/simoninireland/epyc"
