+++ Makefile
@@ -132,7 +132,8 @@ SOURCES_EXTRA = \
 	README.rst \
 	LICENSE \
 	HISTORY \
-	CONTRIBUTING.rst
+	CONTRIBUTING.rst \
+	CITATION.cff
 SOURCES_GENERATED = \
 	MANIFEST \
 	setup.py
