+++ doc/citing.rst
@@ -1,0 +1,18 @@
+Citing
+======
+
+If you use ``epyc`` in your work, find it useful, and want to
+acknowledge it, you can cite the following reference:
+
+Simon Dobson. epyc. `doi:10.5281/zenodo.6460760 <https://github.com/simoninireland/epyc>`_. 2022.
+
+One possible BibTeX record of this is::
+
+  @software{epyc-2022,
+  author = {Simon Dobson},
+  title = {{epyc}},
+  doi = {10.5281/zenodo.6460760},
+  month = {April},
+  url = {https://github.com/simoninireland/epyc},
+  year = {2022},
+  }
